+++ examples/bulk-import/file-import/attributes.php
@@ -6,7 +6,7 @@ try {
     $attributes = provideSampleAttributes();
 
     $handler = $sdk->getBulkImportService()->createFileImport();
-    $attributeHandler = $handler->createAttributeFeed(CATALOG_CODE);
+    $attributeHandler = $handler->createAttributeFeed();
     $attributeHandler->add($attributes[0]);
     $attributeHandler->add($attributes[1]);
     $attributeHandler->end();